Revert "TValidator initial. RC"
This reverts commit cafe271fc1e0d985966d679903c3def1b268e2cd.

+++ TP Olymp/TaskE.py
@@ -1,5 +1,4 @@
 from math import *
-# TAGS: intergal, itegrate, eval
 
 
 def f(e, x):
@@ -24,7 +23,7 @@ def main():
 	with open('input.txt', 'r') as f:
 		data = f.read().splitlines()
 
-	a, b = map(float, data[0].split())
+	a, b = map(int, data[0].split())
 	e = data[1]
 	##
